pyproject.toml
[project]
name = "gym-agent-test"
version = "0.1.0"
description = "Gym Agent với Langchain và Gemini API"
authors = [
    {name = "Truong Nhat Thach",email = "nhatthach2703@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0", 
    "python-dotenv>=1.0.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langchainhub>=0.1.0",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "rich (>=14.2.0,<15.0.0)"
]

[tool.poetry]
packages = [{include = "gym_agent_test", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)"
]
